% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{package}
\name{mitch}
\alias{mitch}
\alias{mitch-package}
\title{Mitch: An R package for multi-dimensional pathway enrichment analysis}
\description{
Mitch is an R package for multi-dimensional enrichment analysis. At it's heart, it uses 
a rank-MANOVA based statistical approach to detect sets of genes that exhibit 
enrichment in the multidimensional space as compared to the background. Mitch is useful 
for pathway analysis of profiling studies with two to or more contrasts, or in studies 
with multiple omics profiling, for example proteomic, transcriptomic, epigenomic 
analysis of the same samples. Mitch is perfectly suited for pathway level differential 
analysis of scRNA-seq data.
}
\details{
A typical Mitch workflow consists of:
 (1) Import gene sets with gmt_import()
 (2) Import profiling data with mitch_import()
 (3) Calculate enrichments with mitch_calc()
 (4) And generate plots and reports with mitch_plots() and mitch_report()

More documentation on the github page https://github.com/markziemann/Mitch or 
 with ?function, eg ?mitch_import
}
