% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{mitch_plots}
\alias{mitch_plots}
\title{mitch_plots}
\usage{
mitch_plots(res, outfile = "Rplots.pdf", cores = detectCores() - 1)
}
\arguments{
\item{res}{a mitch results object.}

\item{outfile}{the destination file for the plots in PDF format. should contain 'pdf' suffix. Defaults to 
'Rplots.pdf'}

\item{cores}{number of parallel cores for plotting}
}
\value{
generates a PDF file containing enrichment plots.
}
\description{
This function generates several plots of multivariate gene set enrichment in high resolution PDF format.
The number of detailed sets to generate is dictated by the resrows set in the mitch_calc command.
}
\examples{
# render enrichment plots in high res pdf
gsets<-gmt_import(system.file('sample_genesets.gmt', package = 'mitch'))
x<-read.table(system.file('rna.rnk', package = 'mitch'),header=TRUE,row.names=1)
res<-mitch_calc(x,gsets,priority='effect',resrows=6,minsetsize=5,cores=2)
mitch_plots(res,outfile='outres.pdf',cores=2)
}
\keyword{mitch}
\keyword{pdf}
\keyword{plot}
\keyword{plots}
